INCS-5657 restructure components into single folder, update imports
