+++ spa/src/components/ExampleComponent/index.tsx
