+++ spa/src/components/Requirements/index.tsx
