+++ spa/src/components/index.tsx
@@ -1,0 +1,7 @@
+import { ExampleComponent } from "./ExampleComponent";
+import { RequirementComponent } from "./Requirements";
+import SyncHeader from "./SyncHeader";
+
+export { ExampleComponent };
+export { RequirementComponent };
+export { SyncHeader };
